Enhancing standalone to not require chroot
I have gone back and forth on this but I think running the agent in as normal
and environment as possible is best.

With these changes, the agent can be deployed to any arbitrary path in the filesystem
and has simple access to the system dirs that it needs to see to pull host metrics
(e.g. /proc, /dev, etc.).

There is some manipulation of LD_LIBRARY_PATH that has to be done, but it is minimal
compared to previous bundles.

This should also make the docker version of the agent and the standalone version as
identical as possible so that we have less to maintain between the two.